+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/npc/citizens/CitizensAdapter.java
@@ -2,28 +2,20 @@ package org.betonquest.betonquest.compatibility.npc.citizens;
 
 import net.citizensnpcs.api.event.SpawnReason;
 import net.citizensnpcs.api.npc.NPC;
-import net.citizensnpcs.trait.HologramTrait;
+import net.citizensnpcs.api.trait.trait.PlayerFilter;
 import org.betonquest.betonquest.api.profile.OnlineProfile;
 import org.betonquest.betonquest.api.quest.npc.Npc;
 import org.bukkit.Location;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.LivingEntity;
 import org.bukkit.event.player.PlayerTeleportEvent;
-import org.bukkit.plugin.Plugin;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
 
 /**
  * Citizens Compatibility Adapter for BetonQuest Npcs.
  */
 public class CitizensAdapter implements Npc<NPC> {
-
-    /**
-     * The plugin instance.
-     */
-    private final Plugin plugin;
 
     /**
      * The Citizens NPC instance.
@@ -33,11 +25,9 @@ public class CitizensAdapter implements Npc<NPC> {
     /**
      * Create a new Citizens Npc Adapter.
      *
-     * @param plugin the plugin instance
-     * @param npc    the Citizens NPC instance
+     * @param npc the Citizens NPC instance
      */
-    public CitizensAdapter(final Plugin plugin, final NPC npc) {
-        this.plugin = plugin;
+    public CitizensAdapter(final NPC npc) {
         this.npc = npc;
     }
 
@@ -104,27 +94,21 @@ public class CitizensAdapter implements Npc<NPC> {
 
     @Override
     public void show(final OnlineProfile onlineProfile) {
-        getEntityList(npc).forEach(entity -> onlineProfile.getPlayer().showEntity(plugin, entity));
+        final PlayerFilter filter = npc.getOrAddTrait(PlayerFilter.class);
+        if (filter.isAllowlist()) {
+            filter.addPlayer(onlineProfile.getPlayerUUID());
+        } else {
+            filter.removePlayer(onlineProfile.getPlayerUUID());
+        }
     }
 
     @Override
     public void hide(final OnlineProfile onlineProfile) {
-        getEntityList(npc).forEach(entity -> onlineProfile.getPlayer().hideEntity(plugin, entity));
-    }
-
-    private List<Entity> getEntityList(final NPC npc) {
-        final List<Entity> entityList = new ArrayList<>();
-        entityList.add(npc.getEntity());
-
-        final HologramTrait hologramTrait = npc.getTraitNullable(HologramTrait.class);
-        if (hologramTrait != null) {
-            final Entity nameEntity = hologramTrait.getNameEntity();
-            if (nameEntity != null) {
-                entityList.add(nameEntity);
-            }
-            entityList.addAll(hologramTrait.getHologramEntities());
+        final PlayerFilter filter = npc.getOrAddTrait(PlayerFilter.class);
+        if (filter.isDenylist()) {
+            filter.addPlayer(onlineProfile.getPlayerUUID());
+        } else {
+            filter.removePlayer(onlineProfile.getPlayerUUID());
         }
-
-        return entityList;
     }
 }