+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/npc/citizens/CitizensWrapper.java
@@ -7,18 +7,12 @@ import org.betonquest.betonquest.api.instruction.Argument;
 import org.betonquest.betonquest.api.profile.Profile;
 import org.betonquest.betonquest.api.quest.npc.Npc;
 import org.betonquest.betonquest.api.quest.npc.NpcWrapper;
-import org.bukkit.plugin.Plugin;
 import org.jetbrains.annotations.Nullable;
 
 /**
  * Citizens wrapper to get a Npc.
  */
 class CitizensWrapper implements NpcWrapper<NPC> {
-
-    /**
-     * The plugin instance.
-     */
-    private final Plugin plugin;
 
     /**
      * Source Registry of NPCs to use.
@@ -33,12 +27,10 @@ class CitizensWrapper implements NpcWrapper<NPC> {
     /**
      * Create a new Citizens Npc Wrapper.
      *
-     * @param plugin   the plugin instance
      * @param registry the registry of NPCs to use
      * @param npcId    the id of the Npc, greater or equals to zero
      */
-    public CitizensWrapper(final Plugin plugin, final NPCRegistry registry, final Argument<Number> npcId) {
-        this.plugin = plugin;
+    public CitizensWrapper(final NPCRegistry registry, final Argument<Number> npcId) {
         this.registry = registry;
         this.npcId = npcId;
     }
@@ -50,6 +42,6 @@ class CitizensWrapper implements NpcWrapper<NPC> {
         if (npc == null) {
             throw new QuestException("NPC with ID " + npcId + " not found");
         }
-        return new CitizensAdapter(plugin, npc);
+        return new CitizensAdapter(npc);
     }
 }